nothing

+++ src/kernel/dev/time.c
@@ -17,7 +17,7 @@ static void pit_init (void) {
     outb(PIT_COMMAND_MODE_PORT, PIT_CHANNLE0 | PIT_LOAD_LOHI | PIT_MODE3);
     outb(PIT_CHANNEL0_DATA_PORT, reload_count & 0xFF);          // 加载低8位
     outb(PIT_CHANNEL0_DATA_PORT, (reload_count >> 8) & 0xFF);   // 再加载高8位
-    irq_install(IRQ0_TIMER, exception_handler_time);    
+    irq_install(IRQ0_TIMER, (uint32_t)exception_handler_time);    
     irq_enable(IRQ0_TIMER);
 }
